requirements.txt
# This file is used to install the required packages for the project.
streamlit
faiss-cpu
sentence-transformers
requests
torch
